backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-docx>=1.1.0
docxtpl>=0.16.8
pyreadstat>=1.2.6
pandas>=2.2.0
openpyxl>=3.1.2
pydantic>=2.6.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
supabase>=2.3.4
scipy>=1.11.0
statsmodels>=0.14.0
